app/agent/requirements.txt
solana>=0.36.0
requests>=2.32.0
cdp-sdk>=0.1.0
python-dotenv>=1.0.0
anchorpy>=0.18.0
solders>=0.18.0
flask>=3.0.0
flask-cors>=4.0.0
